🚑HotFix/Fixed logic to correctly fetch sub breeds
