+++ src/components/DogList.jsx
@@ -1,6 +1,9 @@
 import { useEffect, useState } from "react";
 import { TreeSelect } from "antd";
 import { Card } from "antd";
+import { baseURL, allBreedsURL } from "../utils/endpoints";
+import { handleTreeData } from "../utils/handlers";
+import { gridStyle } from "./styles";
 import axios from "axios";
 
 const DogList = () => {
@@ -8,9 +11,8 @@ const DogList = () => {
   const [valuesByBreed, setValuesByBreed] = useState([]);
 
   useEffect(() => {
-    const url = "https://dog.ceo/api/breeds/list/all";
     axios
-      .get(url)
+      .get(allBreedsURL)
       .then((res) => {
         setAllBreeds(res.data.message);
       })
@@ -26,9 +28,12 @@ const DogList = () => {
     const subBreedValues = [].concat(...Object.values(allBreeds));
     const itsSubBreed = subBreedValues.includes(lastSelectedBreed.toString());
     const breedPics = [];
+
     if (itsSubBreed) {
+      const originalBreed = Object.entries(allBreeds).find(([key, value]) => value.includes(lastSelectedBreed.toString()));
+      const subBreedsEndpoint = `${originalBreed[0]}/${lastSelectedBreed.toString()}/images`
       axios
-        .get(`https://dog.ceo/api/breed/${lastSelectedBreed}/images`)
+        .get(baseURL + subBreedsEndpoint)
         .then((res) => {
           const breedImages = res.data.message;
           for (let image of breedImages) {
@@ -40,9 +45,9 @@ const DogList = () => {
           console.log(err);
         });
     } else {
-      const breedsURL = `https://dog.ceo/api/breed/${lastSelectedBreed}/images`;
+      const breedEndpoint = `${lastSelectedBreed}/images`
       axios
-        .get(breedsURL)
+        .get(baseURL + breedEndpoint)
         .then((res) => {
           const breedImages = res.data.message;
           for (let image of breedImages) {
@@ -56,8 +61,6 @@ const DogList = () => {
     }
   };
 
-  console.log("values", allBreeds);
-
   const { SHOW_PARENT } = TreeSelect;
 
   const tProps = {
@@ -69,34 +72,11 @@ const DogList = () => {
       width: "30%",
     },
   };
-
-  const gridStyle = {
-    width: "20%",
-    padding: "4rem 2rem 4rem 2rem",
-    marginTop: "1rem",
-    marginLeft: "4.5rem",
-    textAlign: "center",
-    backgroundColor: "#DE2151",
-  };
-
-  const handleTreeData = () => {
-    return Object.entries(allBreeds).map(([breed, subBreed], i) => ({
-      title: breed,
-      value: breed,
-      children: subBreed.map((x, i) => {
-        return {
-          title: x,
-          value: x,
-        };
-      }),
-    }));
-  };
-
-  console.log('byBreed', valuesByBreed)
+  
 
   return (
     <>
-      <TreeSelect {...tProps} treeData={handleTreeData()} />
+      <TreeSelect {...tProps} treeData={handleTreeData(allBreeds)} />
       <Card>
         {valuesByBreed.map((x, i) => {
           return (
