+++ src/components/styles.js
@@ -1,0 +1,10 @@
+const gridStyle = {
+    width: "20%",
+    padding: "4rem 2rem 4rem 2rem",
+    marginTop: "1rem",
+    marginLeft: "4.5rem",
+    textAlign: "center",
+    backgroundColor: "#DE2151",
+  };
+
+export { gridStyle }
