+++ src/utils/endpoints.js
@@ -1,0 +1,4 @@
+const allBreedsURL = "https://dog.ceo/api/breeds/list/all";
+const baseURL = "https://dog.ceo/api/breed/"
+
+export { baseURL, allBreedsURL }
